requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
reportlab==4.4.2
python-dateutil==2.9.0.post0

